Prepend Rails.root on Cells.view_paths
__Problem__

Cells use `app/cells` as `view_paths` path and thats ok when you are
using in a gem or rails app that `app/cells` is on the root path. But in
some cases, where rails is not on the root path (like our test app) it's
a problem.

__Solution__

Prepend Rails.root on Cells.view_paths because rails put its root_path
on Rails.root, then when we prepend, Cells will look for files on the
right location.

This has no impact on Engines, because when using Engines is needed to
prepend the Engines root like specified on the
[docs](http://trailblazer.to/gems/cells/rails.html#engine-view-paths).