Add sass-rails to main Gemfile
Without it the specs weren't working. The travis builds in the main repo
were working because it had the gem cached apparently. Attempting to run
a build on a forked repo would fail with the following error:
  LoadError: cannot load such file -- sass

This commit also fixes the tests that check the compilation result of
sass (no whitespaces vs whitespaces)